+++ EssentialFeedTests/Feed Cache/CacheFeedImageDataUseCaseTests.swift
@@ -1,0 +1,38 @@
+//
+//  CacheFeedImageDataUseCaseTests.swift
+//  EssentialFeedTests
+//
+//  Created by Viral on 27/09/22.
+//
+
+import XCTest
+import EssentialFeed
+
+class CacheFeedImageDataUseCaseTests: XCTestCase {
+
+    func test_init_doesNotMessageStoreUponCreation() {
+        let (_ , store) = makeSUT()
+
+        XCTAssertTrue(store.receivedMsg.isEmpty)
+    }
+
+    func test_saveImageDataForURL_requetsImageDataInsertionForURL() {
+        let (sut, store) = makeSUT()
+        let url = anyURL()
+        let data = anyData()
+
+        sut.save(data, for: url) { _ in }
+
+        XCTAssertEqual(store.receivedMsg, [.insert(data: data, for: url)])
+    }
+    // MARK: - Helpers
+
+    private func makeSUT(currentDate: @escaping () -> Date = Date.init, file: StaticString = #file,
+                         line: UInt = #line) -> (sut: LocalFeedImageDataLoader, store: FeedImageDataStoreSpy) {
+        let store = FeedImageDataStoreSpy()
+        let sut = LocalFeedImageDataLoader(store: store)
+        trackForMemoryLeaks(store, file: file, line: line)
+        trackForMemoryLeaks(sut, file: file, line: line)
+        return (sut, store)
+    }
+}
